update scripts, update pbw bundle completed

+++ labs/RuntimeModernization/scripts/build-container.sh
@@ -3,11 +3,32 @@
 ######################
 
 ## No input parameters required to run this script. 
-#clenup studentmig bundle if present.
-#unzip completed mig bundle from cloned git repo to student folder
+## -i  If you want to run the commands in the script by being prompted to continue
+## exmple: build-container.sh -i
+
+numParms=$#
+parm1=$1
+
+#echo "numparms: $numParms"
+#echo "parm 1: $1"
+
+if [ $numParms != 0 ]; then
+#  echo "numParms > 0, have parms"
+  if [ $1 == "-i" ]; then
+#    echo "parm1 is '-i"
+    INTERACTIVE_MODE="true"
+    echo ""
+    echo "---------------------------------------------"
+    echo "    Script is running in interactive mode! "
+    echo "---------------------------------------------"
+    echo ""
+    sleep 5
+  fi 
+fi
+
 
 STUDENT_LAB_DIR="/home/techzone/Student/labs/appmod"
-STUDENT_PBW_BUNDLE="/home/techzone/Student/labs/appmod/plantsbywebsphereee6.ear_migrationBundle"
+STUDENT_PBW_BUNDLE="/home/techzone/Student/labs/appmod/pbw-bundle-complete"
 PWD=pwd
 LOGS=$STUDENT_LAB_DIR/logs
 LOG=$STUDENT_LAB_DIR/logs/build-container.log
@@ -76,7 +97,7 @@ if [ -n "$cmd1" ]; then
   echo "   ->  pbw docker image removed"
   sleep 2
 fi
-echo ""
+
 echo "----------------------"
 echo "End of cleanup steps"
 echo "----------------------"
@@ -92,89 +113,132 @@ echo "Setup, Build, and Run the PBW container" | tee -a $LOG
 echo "=======================================" | tee -a $LOG
 echo ""
 
-echo "-> start PBW database in container"
+echo ""
+echo "==========================================="
+echo "1. Start PBW database in container"
 echo "--------------------------"
 echo " " | tee -a $LOG
-echo "   1. -> docker start db2_demo_data" | tee -a $LOG
+echo "   1. ---> docker start db2_demo_data" | tee -a $LOG
 echo " " | tee -a $LOG
 echo "--------------------------"
 echo ""
 
+if [[ $INTERACTIVE_MODE == "true" ]]; then
+  read -n 1 -r -s -p $'Press enter to continue...\n'
+  echo ""
+fi    
 docker start db2_demo_data
 sleep 2
 
 
 echo ""
-echo "-> Verify database container is running"
+echo "==========================================="
+echo "2. Verify database container is running"
 echo "--------------------------"
 echo " " | tee -a $LOG
-echo "   2. -> docker ps | grep db2_demo_data" | tee -a $LOG
+echo "   2. ---> docker ps | grep db2_demo_data" | tee -a $LOG
 echo " " | tee -a $LOG
 echo "--------------------------"
 echo ""
+
+if [[ $INTERACTIVE_MODE == "true" ]]; then
+  read -n 1 -r -s -p $'Press enter to continue...\n'
+  echo ""
+fi    
 docker ps | grep db2_demo_data
 sleep 2
 
 echo ""
-echo "-> Create the docker network, pbw-network"
+echo "==========================================="
+echo "3. Create the docker network, pbw-network"
 echo "---------------------------------"
 echo " " | tee -a $LOG
-echo "   3. -> docker network create pbw-network" | tee -a $LOG
+echo "   3. ---> docker network create pbw-network" | tee -a $LOG
 echo " " | tee -a $LOG
 echo "---------------------------------"
 echo ""
+
+if [[ $INTERACTIVE_MODE == "true" ]]; then
+  read -n 1 -r -s -p $'Press enter to continue...\n'
+  echo ""
+fi    
 docker network create pbw-network
 sleep 2
 
 
 echo ""
-echo "-> Verify docker network is created"
+echo "==========================================="
+echo "4. Verify docker network is created"
 echo "---------------------------------"
 echo " " | tee -a $LOG
-echo "   4. -> docker network list | grep pbw-network" | tee -a $LOG
+echo "   4. ---> docker network list | grep pbw-network" | tee -a $LOG
 echo " " | tee -a $LOG
 echo "---------------------------------"
 echo ""
+
+if [[ $INTERACTIVE_MODE == "true" ]]; then
+  read -n 1 -r -s -p $'Press enter to continue...\n'
+  echo ""
+fi    
 docker network list | grep pbw-network
 
 
 echo ""
-echo "-> Connect database container to the Docker network"
+echo "==========================================="
+echo "5. Connect database container to the Docker network"
 echo "------------------------------------------------"
 echo " " | tee -a $LOG
-echo "   5. -> docker network connect pbw-network db2_demo_data" | tee -a $LOG
+echo "   5. ---> docker network connect pbw-network db2_demo_data" | tee -a $LOG
 echo " " | tee -a $LOG
 echo "------------------------------------------------"
 echo ""
+
+if [[ $INTERACTIVE_MODE == "true" ]]; then
+  read -n 1 -r -s -p $'Press enter to continue...\n'
+  echo ""
+fi    
 docker network connect pbw-network db2_demo_data
 sleep 2
 
 
 
 echo ""
-echo "-> Build and tag the 'PlantsByWebSphere' application container image"
+echo "==========================================="
+echo "6. Build and tag the 'PlantsByWebSphere' application container image"
 echo "--------------------------------------------------"
 echo " " | tee -a $LOG
-echo "   6. -> docker build -f /home/techzone/Student/labs/appmod/plantsbywebsphereee6.ear_migrationBundle/Containerfile --tag default-route-openshift-image-registry.apps.ocp.ibm.edu/apps/pbw ." | tee -a $LOG
+echo "   6. ---> docker build -f $STUDENT_PBW_BUNDLE/Containerfile --tag default-route-openshift-image-registry.apps.ocp.ibm.edu/apps/pbw ." | tee -a $LOG
 echo " " | tee -a $LOG
  echo "--------------------------------------------------"
 echo ""
-docker build -f /home/techzone/Student/labs/appmod/plantsbywebsphereee6.ear_migrationBundle/Containerfile --tag default-route-openshift-image-registry.apps.ocp.ibm.edu/apps/pbw .
+
+
+if [[ $INTERACTIVE_MODE == "true" ]]; then
+  read -n 1 -r -s -p $'Press enter to continue...\n'
+  echo ""
+fi    
+docker build -f $STUDENT_PBW_BUNDLE/Containerfile --tag default-route-openshift-image-registry.apps.ocp.ibm.edu/apps/pbw .
 sleep 2 
 
 
 echo ""
-echo "-> Run 'PlantsByWebSphere' application in container"
+echo "==========================================="
+echo "7. Run 'PlantsByWebSphere' application in container"
 echo "--------------------------------------------------"
 echo " " | tee -a $LOG
-echo "   7. -> docker run -d --rm -p 9080:9080 -p 9443:9443 --network pbw-network --name pbw  default-route-openshift-image-registry.apps.ocp.ibm.edu/apps/pbw" | tee -a $LOG
+echo "   7. ---> docker run -d --rm -p 9080:9080 -p 9443:9443 --network pbw-network --name pbw  default-route-openshift-image-registry.apps.ocp.ibm.edu/apps/pbw" | tee -a $LOG
 echo " " | tee -a $LOG
 echo "--------------------------------------------------"
 echo ""
+
+if [[ $INTERACTIVE_MODE == "true" ]]; then
+  read -n 1 -r -s -p $'Press enter to continue...\n'
+  echo ""
+fi    
 docker run -d --rm -p 9080:9080 -p 9443:9443 --network pbw-network --name pbw  default-route-openshift-image-registry.apps.ocp.ibm.edu/apps/pbw
 
 echo ""
-echo "-> pbw app should be running (UP) now!"
+echo "---> pbw app should be running (UP) now!"
 echo "--------------------------------------"
 echo "docker ps | grep pbw" 
 echo "--------------------------------------"
@@ -188,4 +252,3 @@ echo "================================" | tee -a $LOG
 echo "" | tee -a $LOG
 
 exit 0
-
